+++ tests-e2e/welcome.spec.ts
@@ -18,4 +18,12 @@ test("can be closed", async ({ page }) => {
   await expect(page.getByTestId("welcome-description")).not.toBeVisible();
 });
 
-test.skip("it will only show once", async () => {});
+test("it will only show once", async ({ page }) => {
+  await page.getByTestId("get-started-button").click();
+  await expect(page.getByTestId("welcome-description")).not.toBeVisible();
+
+  const target = process.env.CI === "true" ? "/kanbeasy" : "/";
+  await page.goto(target);
+
+  await expect(page.getByTestId("welcome-description")).not.toBeVisible();
+});